Speed up sieve finalize time
Converted `active_m_i` to bitset.
Takes constant `M_INC/2`, instead of old `6*M_INC` space
Sieve finalization takes half as long now!

Overall maybe 1-2% boost in speed from less "wait done X" and
"wait done M"

+++ gpu_sieve.h
@@ -34,7 +34,7 @@ class GPUSieve {
         GPUSieve(const struct Config& config);
         ~GPUSieve();
 
-        uint8_t* run(const uint64_t m_start, const uint64_t m_inc, const uint64_t X, const uint32_t max_p_i);
+        uint64_t* run(const uint64_t m_start, const uint64_t m_inc, const uint64_t X, const uint32_t max_p_i);
 
     private:
         cudaStream_t runner;
@@ -74,14 +74,14 @@ class GPUSieve {
 
         size_t composite_bytes;
         uint8_t *composite;
-        uint8_t *composite_compressed;
+        uint64_t *composite_compressed;
 
         uint8_t *D_wheel;
         /******************************/
         /******** GPU POINTERS ********/
 
         // Host side
-        uint8_t* host_composite;
+        uint64_t* host_composite;
         int64_t  *host_thread_stats;
         vector<std::pair<uint32_t, uint32_t>> d_neg_inv_K;
         vector<uint8_t> host_wheel;
